requirements.txt
Flask>=0.10.1
jsonresume-validator>=0.1.0
colander
jinja2